# Use SSE 4.2 due to parse escaped string