@Watson1978 2022-07-09
> I think I should have added the test code...

Added.